django/requirements.txt
attrdict==2.0.1
biopython==1.73
Django==2.2.4
mysqlclient==1.4.2.post1
numpy==1.16.4
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
gunicorn
djangorestframework==3.11.0
django-cors-headers==3.2.1
drf-extensions==0.6.0
python-memcached>=1.3
